mock.waitUntilCalled, close #197
