+++ lib/models/request_mock.js
@@ -3,7 +3,7 @@
 const EventEmitter = require('events');
 const URL = require('url').URL;
 
-const {FatalError, AssertionError} = require('../errors');
+const {FatalError, AssertionError, TimeoutError} = require('../errors');
 const utils = require('../utils');
 
 
@@ -59,8 +59,26 @@ module.exports = class RequestMock {
         this._events.emit("respond");
     }
 
-    increaseCalled() {
+    waitUntilCalled(timeout = 500) {
+        if (this.called) return Promise.resolve();
+        return new Promise((resolve, reject) => {
+            let rejected = false;
+            const tid = setTimeout(() => {
+                rejected = true;
+                reject(new TimeoutError(`Wait until mock of "${this.url}" is called`, timeout));
+            }, timeout);
+            this._events.once("on-request", () => {
+                if (!rejected) {
+                    clearTimeout(tid);
+                    resolve();
+                }
+            });
+        });
+    }
+
+    onRequest() {
         this._timesCalled++;
+        this._events.emit("on-request");
     }
 
     _assertCalled(times, msg) {
