+++ lib/modules/browser_requests.js
@@ -51,6 +51,7 @@ module.exports = class BrowserRequests extends BrowserModule {
     }
 
     _respondWithMock(request, mock) {
+        mock.onRequest();
         if (mock.auto && mock.immediate) {
             return request.respond(mock.response);
         } else if (mock.auto) {
