+++ lib/request_mocker.js
@@ -13,14 +13,10 @@ module.exports = class RequestMocker {
     getMockedResponse(request) {
         const url = new URL(request.url());
         const method = request.method();
-        const mock = this._getMock(`${url.origin}${url.pathname}`, {
+        return this._getMock(`${url.origin}${url.pathname}`, {
             method: method,
             queryString: url.search ? utils.parseQueryString(url) : undefined
         });
-        if (mock) { // TODO: move this to request ?
-            mock.increaseCalled();
-            return mock;
-        }
     }
 
     mockRequest(url, options) {
